requirements-dev.txt
-e .  # Install the production requirements

pytest==2.7.1
pytest-django==2.8.0